Move media-library-tools out of the framework
Second pure move, same reason as brand-context-tools:
`src/lib/agent/tools/media-library-tools.ts` becomes
`src/lib/server/media-library-tools.ts`.

79 lines that import `ai`, `zod`, a Supabase type and `brand-media` —
nothing from the framework — but living in the directory that is about
to be deleted, while Motion Video and the UGC producer depend on them to
look at and use the brand's media library.

Motion Video and the framework's own two barrels follow the path here.
The two UGC files pick it up in the next commit, where their loops move
too, because their import and their rewrite touch the same lines.
